Tone down statement
"W3C is handling this issue with a Threat Model." implies endorsement by W3C members. But the link brings the reader to a use case within the document, which attempts to summarize an introduction to "Trust over IP" from 2021 by the "trust over IP foundation" which is a 35-page PDF hosted elsewhere. 

This is misleading and risky to make such a claim and to rely on a document whose status is unclear at a cursory glance.